Read a Wavefront OBJ 3D scene file into an R list

Read a Wavefront OBJ 3D scene file into an R list

read.obj(f, materialspath = NULL, convert.rgl = FALSE)

Arguments

f
Path to an OBJ file

materialspath
Path to a folder containing materials files. This is optional and only required if materials files are in a different folder from the OBJ file defined by f.

convert.rgl
Whether to convert the returned list to a rgl::shapelist3d object containing rgl::mesh3d objects.

Value

When convert.rgl=FALSE, the default, a named list with items shapes and materials, each containing sublists with one entry per object (shapes) or material (materials). Objects in the shapes list have the following structure

positions 3xN set of 3D vertices

normals 3xN set of normal directions for each vertex (has 3 rows and 0 cols when normals are not available)

texcoords vector containing unprocessed texture coordinates

indices 3/4xM set of indices into vertex array (trimesh/quadmesh) 0-indexed

material_ids (0-indexed, -1 when not set)

When convert.rgl=FALSE a list of class shapelist3d containing a mesh3d for each object or group element in the original OBJ file. See tinyobj2shapelist3d for details of rgl conversion.

Sample files

Note that at the request of the CRAN maintainers the sample files have the file extension .wavefront instead of the standard .obj because this triggers a false positive R CMD check NOTE.

See also

Examples

cube=read.obj(system.file("obj/cube.wavefront", package = "readobj"))
str(cube, max.level = 3)

#> List of 2
#>  $ shapes   :List of 6
#>   ..$ front :List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 0 2 2 0 0 2 2 0 2 2 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 0 0
#>   ..$ back  :List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 2 2 0 2 0 0 0 0 0 0 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 0 0
#>   ..$ right :List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 2 2 2 2 0 2 2 0 0 2 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 1 1
#>   ..$ top   :List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 0 2 0 0 2 2 2 2 2 2 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 0 0
#>   ..$ left  :List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 0 2 0 0 0 0 0 0 2 0 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 2 2
#>   ..$ bottom:List of 5
#>   .. ..$ positions   : num [1:3, 1:4] 0 0 2 0 0 0 2 0 0 2 ...
#>   .. ..$ normals     : num[1:3, 0 ] 
#>   .. ..$ texcoords   : num(0) 
#>   .. ..$ indices     : num [1:3, 1:2] 0 1 2 0 2 3
#>   .. ..$ material_ids: int [1:2] 0 0
#>  $ materials:List of 5
#>   ..$ white:List of 13
#>   .. ..$ ambient         : num [1:3] 0 0 0
#>   .. ..$ diffuse         : num [1:3] 1 1 1
#>   .. ..$ specular        : num [1:3] 0 0 0
#>   .. ..$ transmittance   : num [1:3] 0 0 0
#>   .. ..$ emission        : num [1:3] 0 0 0
#>   .. ..$ shininess       : num 1
#>   .. ..$ ior             : num 1
#>   .. ..$ dissolve        : num 1
#>   .. ..$ illum           : int 0
#>   .. ..$ ambient_texname : chr ""
#>   .. ..$ diffuse_texname : chr ""
#>   .. ..$ specular_texname: chr ""
#>   .. ..$ normal_texname  : chr ""
#>   ..$ red  :List of 13
#>   .. ..$ ambient         : num [1:3] 0 0 0
#>   .. ..$ diffuse         : num [1:3] 1 0 0
#>   .. ..$ specular        : num [1:3] 0 0 0
#>   .. ..$ transmittance   : num [1:3] 0 0 0
#>   .. ..$ emission        : num [1:3] 0 0 0
#>   .. ..$ shininess       : num 1
#>   .. ..$ ior             : num 1
#>   .. ..$ dissolve        : num 1
#>   .. ..$ illum           : int 0
#>   .. ..$ ambient_texname : chr ""
#>   .. ..$ diffuse_texname : chr ""
#>   .. ..$ specular_texname: chr ""
#>   .. ..$ normal_texname  : chr ""
#>   ..$ green:List of 13
#>   .. ..$ ambient         : num [1:3] 0 0 0
#>   .. ..$ diffuse         : num [1:3] 0 1 0
#>   .. ..$ specular        : num [1:3] 0 0 0
#>   .. ..$ transmittance   : num [1:3] 0 0 0
#>   .. ..$ emission        : num [1:3] 0 0 0
#>   .. ..$ shininess       : num 1
#>   .. ..$ ior             : num 1
#>   .. ..$ dissolve        : num 1
#>   .. ..$ illum           : int 0
#>   .. ..$ ambient_texname : chr ""
#>   .. ..$ diffuse_texname : chr ""
#>   .. ..$ specular_texname: chr ""
#>   .. ..$ normal_texname  : chr ""
#>   ..$ blue :List of 13
#>   .. ..$ ambient         : num [1:3] 0 0 0
#>   .. ..$ diffuse         : num [1:3] 0 0 1
#>   .. ..$ specular        : num [1:3] 0 0 0
#>   .. ..$ transmittance   : num [1:3] 0 0 0
#>   .. ..$ emission        : num [1:3] 0 0 0
#>   .. ..$ shininess       : num 1
#>   .. ..$ ior             : num 1
#>   .. ..$ dissolve        : num 1
#>   .. ..$ illum           : int 0
#>   .. ..$ ambient_texname : chr ""
#>   .. ..$ diffuse_texname : chr ""
#>   .. ..$ specular_texname: chr ""
#>   .. ..$ normal_texname  : chr ""
#>   ..$ light:List of 13
#>   .. ..$ ambient         : num [1:3] 1 1 1
#>   .. ..$ diffuse         : num [1:3] 1 1 1
#>   .. ..$ specular        : num [1:3] 0 0 0
#>   .. ..$ transmittance   : num [1:3] 0 0 0
#>   .. ..$ emission        : num [1:3] 0 0 0
#>   .. ..$ shininess       : num 1
#>   .. ..$ ior             : num 1
#>   .. ..$ dissolve        : num 1
#>   .. ..$ illum           : int 0
#>   .. ..$ ambient_texname : chr ""
#>   .. ..$ diffuse_texname : chr ""
#>   .. ..$ specular_texname: chr ""
#>   .. ..$ normal_texname  : chr ""

# demonstrate direct conversion of result to rgl format
if(require('rgl')) {
  cuber=read.obj(system.file("obj/cube.wavefront", package = "readobj"),
    convert.rgl=TRUE)
  shade3d(cuber)
}

#> Loading required package: rgl