Convert the raw tinyobjloader shapes/materials list into an rgl shapelist3d

Convert the raw tinyobjloader shapes/materials list into an rgl shapelist3d

tinyobj2shapelist3d(x)

Arguments

x
A raw tinyobjloader shapes/materials list

Value

a list of class shapelist3d containing a mesh3d for each object or group element in the original OBJ file.

Details

Not all materials settings can be processed at the moment. In particular only the following are used:

diffuse -> mapped onto rgl material color field

ambient

specular

emission

See also

read.obj, mesh3d, shapelist3d, rgl.material

Examples

cube=read.obj(system.file("obj/cube.obj", package = "readobj"))
if(require("rgl")){
  cubesl=tinyobj2shapelist3d(cube)
  shade3d(cubesl)
}